+++ scripts/deploy.sh
@@ -4,13 +4,6 @@ set -o errexit
 set -o pipefail
 set -o nounset
 
-# Signal handler
-function cleanup() {
-  perl -i -pe 's/^exclude = \["ubo_app\/services\/\*-speech-synthesis\/models\/\*", (.*)\]$/exclude = [\1]/' pyproject.toml
-}
-trap cleanup ERR
-trap cleanup EXIT
-
 echo $0
 
 deps=${deps:-"False"}
@@ -19,9 +12,7 @@ kill=${kill:-"False"}
 restart=${restart:-"False"}
 env=${env:-"True"}
 
-perl -i -pe 's/^exclude = \[(.*)\]$/exclude = ["ubo_app\/services\/*-speech-synthesis\/models\/*", \1]/' pyproject.toml
 uv build
-cleanup
 LATEST_VERSION=$(basename $(ls -rt dist/*.whl | tail -n 1))
 
 function run_on_pod() {
@@ -51,10 +42,7 @@ function run_on_pod_as_root() {
 scp dist/$LATEST_VERSION ubo-development-pod-$index:/tmp/
 
 run_on_pod "$(if [ "$deps" == "True" ]; then echo "pip install --upgrade /tmp/$LATEST_VERSION &&"; fi)
-mv /opt/ubo/env/lib/python3.*/site-packages/ubo_app/services/*-speech-synthesis/models /tmp/
 pip install --no-index --upgrade --force-reinstal --no-deps /tmp/$LATEST_VERSION
-[ -e /opt/ubo/env/bin/ipdb3 ] || pip install ipdb
-mv /tmp/models /opt/ubo/env/lib/python3.*/site-packages/ubo_app/services/*-speech-synthesis/
 true"
 
 if [ "$bootstrap" == "True" ] || [ "$env" == "True" ] || [ "$restart" == "True" ]; then